backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
supabase
moviepy
openai-whisper
keybert
sentence-transformers
nltk
